+++ config.prod-eu.ts
@@ -1,0 +1,9 @@
+export const API_V2_URL = 'https://api.eu.arena.im/v2';
+export const CACHED_API = 'https://cached-api.eu.arena.im/v1';
+export const DEFAULT_AUTH_TOKEN = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJfaWQiOiI1NGQ5OGJiNmY3MDIyOGU4MWI4Njc5YmUiLCJyb2xlcyI6WyJVU0VSIl0sImV4cCI6MzM2OTQxODM2OSwiaWF0IjoxNDc3MjU4MzY5fQ.dNpdrs3ehrGAhnPFIlWMrQFR4mCFKZl_Lvpxk1Ddp4o';
+export const FIREBASE_APIKEY = 'AIzaSyCdnt__-0U00CfHQu97gzAfub6j7rS3lQc';
+export const FIREBASE_AUTHDOMAIN = 'arena-eu-im-prd.firebaseapp.com';
+export const FIREBASE_PROJECT_ID = 'arena-eu-im-prd';
+export const GRAPHQL_ENDPOINT = 'https://public.eu.arena.im/graphql';
+export const ARENA_REALTIME_URL = 'https://realtime.eu.arena.im';
+export const ARENA_URL = 'https://go.eu.arena.im';